Title: 2개씩 그룹짓기 2, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 250112/2개씩 그룹짓기 2/group-of-pairs-2.java
@@ -5,7 +5,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
 
         int n = sc.nextInt();
-        List<Integer> lst = new ArrayList<>();
+        List<Long> lst = new ArrayList<>();
 
         for (int i = 0; i < n; i++)
             lst.add(sc.nextInt());
